library_backend/requirements.txt
fastapi
gunicorn
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic

aiofiles
cloudinary
python-multipart